More changes to try to make it work.
In fact, this is a dead-end. First, the absence of a FSM makes things
quite hard. Second, the use of several splice related events seems to be
useless: once we cannot splice anymore, we have to disable the events on
both ends, hence we should be able to get the same behaviour with only
events at local and remote endpoints. Last, the inability to know why we
would block when splicing makes things harder.